Move cache types into separate types.py module
Addresses review feedback: split dataclass types from helper
functions in cache.py for clearer module organization.
