+++ claas/training/engine/local/cache.py
@@ -1,43 +1,18 @@
-"""Typed cache structures and helpers for CPU-resident LoRA state between training steps."""
+"""CPU-resident optimizer state helpers for the local training engine."""
 
 from __future__ import annotations
 
 import copy
-from dataclasses import dataclass
 from typing import cast
 
 import torch
 
-from claas.core.types import DistillResponse
-
-
-@dataclass(frozen=True, slots=True)
-class LoraAdapterConfig:
-    """Typed representation of LoRA adapter configuration."""
-
-    r: int
-    lora_alpha: int
-    target_modules: list[str]
-    lora_dropout: float
-    bias: str
-    task_type: str
-
-
-@dataclass(frozen=True, slots=True)
-class LoraCacheEntry:
-    """CPU-resident snapshot of LoRA adapter state between training steps."""
-
-    lora_state_dict: dict[str, torch.Tensor]
-    optimizer_state_dict: dict[str, object]
-    adapter_config: LoraAdapterConfig
-
-
-@dataclass(frozen=True, slots=True)
-class DistillStepResult:
-    """Result of a distillation step with both response and cache entry."""
-
-    response: DistillResponse
-    cache_entry: LoraCacheEntry
+# Re-export types for backward compatibility
+from claas.training.engine.local.types import (  # noqa: F401
+    DistillStepResult,
+    LoraAdapterConfig,
+    LoraCacheEntry,
+)
 
 
 def cpu_optimizer_state(state_dict: dict[str, object]) -> dict[str, object]:
